Move feedback model to db.model
